move "hir typeck" into the `type_check` module
It is just methods that are generic over the family F.
