+++ src/hir.rs
@@ -9,7 +9,6 @@ use crate::ty::declaration::Declaration;
 use std::sync::Arc;
 
 crate mod query_definitions;
-crate mod type_check;
 
 salsa::query_group! {
     crate trait HirDatabase: salsa::Database {
@@ -220,5 +219,5 @@ index_type! {
 
 #[derive(Copy, Clone, Debug, PartialEq, Eq, Hash)]
 crate struct IdentifierData {
-    text: StringId,
+    crate text: StringId,
 }
